testing for type name.
