+++ src/modules/connector-hub/components/card-collection/__tests__/connector-card.test.tsx
@@ -2,8 +2,9 @@ import { mount } from 'enzyme';
 import React from 'react';
 import { ConnectorCard } from '../connector-card';
 import { StyledCard, StyledMeta } from '../connector-card.css';
-import { Row, Col } from 'antd';
+import { Row, Col, Tag, Card } from 'antd';
 import { ConnectorModal } from '../../../../../redux/connector-hub/types';
+import { TYPE_NAMES_DATA } from '../../../../../fixtures/connector-type-data';
 
 describe('ConnectorCard', () => {
   const connector = (overrides: object = {}): ConnectorModal => ({
@@ -49,4 +50,12 @@ describe('ConnectorCard', () => {
     expect(columns.at(0).text()).toEqual('ERN');
     expect(columns.at(1).text()).toEqual(connector().ern);
   });
+
+  it('renders a tag with typeId', () => {
+    const tags = this.connectorCard.find(Tag);
+    expect(this.connectorCard.find(Card)).toHaveLength(1);
+    expect(tags).toHaveLength(1);
+    expect(tags.at(0).text()).toEqual(TYPE_NAMES_DATA.find(myObj => myObj.id === connector().typeId).name);
+  });
+  
 });
